requirements/common.txt
django==3.0.4
djangorestframework
djangorestframework-jwt
django-cors-headers
django-pydenticon
pillow
black